Change Earth-based to Ground-based. Implement CORSS. Refactor common populate functions.
