+++ opus/import/populate_obs_instrument_COUVIS.py
@@ -17,6 +17,7 @@ import impglobals
 import import_util
 
 from populate_obs_mission_cassini import *
+from populate_util import *
 
 
 ################################################################################
@@ -124,57 +125,16 @@ def populate_obs_general_COUVIS_observation_type_OBS(**kwargs):
     return 'SCU' # Spectral Cube
 
 def populate_obs_general_COUVIS_time1_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    index_row = metadata['index_row']
-    start_time = import_util.safe_column(index_row, 'START_TIME')
-
-    if start_time is None:
-        return None
-
-    try:
-        start_time_sec = julian.tai_from_iso(start_time)
-    except Exception as e:
-        import_util.log_nonrepeating_warning(
-            f'Bad start time format "{start_time}": {e}')
-        return None
-
-    return start_time_sec
+    return populate_time1_from_index(**kwargs)
 
 def populate_obs_general_COUVIS_time2_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    index_row = metadata['index_row']
-    stop_time = import_util.safe_column(index_row, 'STOP_TIME')
-
-    if stop_time is None:
-        return None
-
-    try:
-        stop_time_sec = julian.tai_from_iso(stop_time)
-    except Exception as e:
-        import_util.log_nonrepeating_warning(
-            f'Bad stop time format "{stop_time}": {e}')
-        return None
-
-    general_row = metadata['obs_general_row']
-    start_time_sec = general_row['time1']
-
-    if start_time_sec is not None and stop_time_sec < start_time_sec:
-        start_time = import_util.safe_column(index_row, 'START_TIME')
-        import_util.log_warning(f'time1 ({start_time}) and time2 ({stop_time}) '
-                                f'are in the wrong order - setting to time1')
-        stop_time_sec = start_time_sec
-
-    return stop_time_sec
+    return populate_time2_from_index(**kwargs)
 
 def populate_obs_general_COUVIS_target_name_OBS(**kwargs):
     return helper_cassini_intended_target_name(**kwargs)
 
 def populate_obs_general_COUVIS_observation_duration_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    obs_general_row = metadata['obs_general_row']
-    time_sec1 = obs_general_row['time1']
-    time_sec2 = obs_general_row['time2']
-    return max(time_sec2 - time_sec1, 0)
+    return populate_observation_duration_from_time(**kwargs)
 
 def populate_obs_pds_COUVIS_note_OBS(**kwargs):
     metadata = kwargs['metadata']
@@ -193,36 +153,15 @@ def populate_obs_pds_COUVIS_primary_file_spec_OBS(**kwargs):
     return _COUVIS_file_spec_helper(**kwargs)
 
 def populate_obs_pds_COUVIS_product_creation_time_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    index_label = metadata['index_label']
-    pct = index_label['PRODUCT_CREATION_TIME']
-
-    try:
-        pct_sec = julian.tai_from_iso(pct)
-    except Exception as e:
-        import_util.log_nonrepeating_warning(
-            f'Bad product creation time format "{pct}": {e}')
-        return None
-
-    return pct_sec
+    return populate_product_creation_time_from_index_label(**kwargs)
 
 # Format: "CO-S-UVIS-2-SSB-V1.4"
 def populate_obs_pds_COUVIS_data_set_id_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    index_row = metadata['index_row']
-    dsi = index_row['DATA_SET_ID']
-
-    return (dsi, dsi)
+    return populate_data_set_id_from_index(**kwargs)
 
 # Format: "EUV2015_001_17_57"
 def populate_obs_pds_COUVIS_product_id_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    supp_index_row = metadata.get('supp_index_row', None)
-    if supp_index_row is None:
-        return None
-    product_id = supp_index_row['PRODUCT_ID']
-
-    return product_id
+    return populate_product_id_from_supp_index(**kwargs)
 
 # We occasionally don't bother to generate ring_geo data for COUVIS, like during
 # cruise, so just use the given RA/DEC from the label if needed. We don't make
@@ -266,12 +205,6 @@ def populate_obs_general_COUVIS_declination2_OBS(**kwargs):
     index_row = metadata['index_row']
     dec = import_util.safe_column(index_row, 'DECLINATION')
     return dec
-
-def populate_obs_mission_cassini_COUVIS_mission_phase_name_OBS(**kwargs):
-    return helper_cassini_mission_phase_name(**kwargs)
-
-def populate_obs_mission_cassini_COUVIS_sequence_id_OBS(**kwargs):
-    return None
 
 
 ### OBS_TYPE_IMAGE TABLE ###
@@ -616,6 +549,13 @@ def populate_obs_mission_cassini_COUVIS_spacecraft_clock_count2_OBS(**kwargs):
         return None
     new_count_sec = count_sec + (time2-time1)
     return new_count_sec
+
+def populate_obs_mission_cassini_COUVIS_mission_phase_name_OBS(**kwargs):
+    return helper_cassini_mission_phase_name(**kwargs)
+
+def populate_obs_mission_cassini_COUVIS_sequence_id_OBS(**kwargs):
+    return None
+
 
 ################################################################################
 # THESE ARE SPECIFIC TO OBS_INSTRUMENT_COUVIS
