+++ opus/import/populate_obs_instrument_COVIMS.py
@@ -17,6 +17,7 @@ import impglobals
 import import_util
 
 from populate_obs_mission_cassini import *
+from populate_util import *
 
 
 ################################################################################
@@ -91,57 +92,16 @@ def populate_obs_general_COVIMS_observation_type_OBS(**kwargs):
     return 'SCU' # Spectral Cube
 
 def populate_obs_general_COVIMS_time1_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    index_row = metadata['index_row']
-    start_time = index_row['START_TIME']
-
-    if start_time is None:
-        return None
-
-    try:
-        start_time_sec = julian.tai_from_iso(start_time)
-    except Exception as e:
-        import_util.log_nonrepeating_error(
-            f'Bad start time format "{start_time}": {e}')
-        return None
-
-    return start_time_sec
+    return populate_time1_from_index(**kwargs)
 
 def populate_obs_general_COVIMS_time2_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    index_row = metadata['index_row']
-    stop_time = import_util.safe_column(index_row, 'STOP_TIME')
-
-    if stop_time is None:
-        return None
-
-    try:
-        stop_time_sec = julian.tai_from_iso(stop_time)
-    except Exception as e:
-        import_util.log_nonrepeating_error(
-            f'Bad stop time format "{stop_time}": {e}')
-        return None
-
-    general_row = metadata['obs_general_row']
-    start_time_sec = general_row['time1']
-
-    if start_time_sec is not None and stop_time_sec < start_time_sec:
-        start_time = import_util.safe_column(index_row, 'START_TIME')
-        import_util.log_warning(f'time1 ({start_time}) and time2 ({stop_time}) '
-                                f'are in the wrong order - setting to time1')
-        stop_time_sec = start_time_sec
-
-    return stop_time_sec
+    return populate_time2_from_index(**kwargs)
 
 def populate_obs_general_COVIMS_target_name_OBS(**kwargs):
     return helper_cassini_intended_target_name(**kwargs)
 
 def populate_obs_general_COVIMS_observation_duration_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    obs_general_row = metadata['obs_general_row']
-    time_sec1 = obs_general_row['time1']
-    time_sec2 = obs_general_row['time2']
-    return max(time_sec2 - time_sec1, 0)
+    return populate_observation_duration_from_time(**kwargs)
 
 def populate_obs_pds_COVIMS_note_OBS(**kwargs):
     None
@@ -153,35 +113,15 @@ def populate_obs_pds_COVIMS_primary_file_spec_OBS(**kwargs):
     return _COVIMS_file_spec_helper(**kwargs)
 
 def populate_obs_pds_COVIMS_product_creation_time_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    index_label = metadata['index_label']
-    pct = index_label['PRODUCT_CREATION_TIME']
-
-    try:
-        pct_sec = julian.tai_from_iso(pct)
-    except Exception as e:
-        import_util.log_nonrepeating_error(
-            f'Bad product creation time format "{pct}": {e}')
-        return None
-
-    return pct_sec
+    return populate_product_creation_time_from_index_label(**kwargs)
 
 # Format: "CO-E/V/J/S-VIMS-2-QUBE-V1.0"
 def populate_obs_pds_COVIMS_data_set_id_OBS(**kwargs):
-    # For VIMS the DATA_SET_ID is provided in the volume label file,
-    # not the individual observation rows
-    metadata = kwargs['metadata']
-    index_label = metadata['index_label']
-    dsi = index_label['DATA_SET_ID']
-    return (dsi, dsi)
+    return populate_data_set_id_from_index_label(**kwargs)
 
 # Format: "1/1294638283_1"
 def populate_obs_pds_COVIMS_product_id_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    index_row = metadata['index_row']
-    product_id = index_row['PRODUCT_ID']
-
-    return product_id
+    return populate_product_id_from_index(**kwargs)
 
 def populate_obs_general_COVIMS_right_asc1_OBS(**kwargs):
     metadata = kwargs['metadata']
@@ -222,15 +162,6 @@ def populate_obs_general_COVIMS_declination2_OBS(**kwargs):
     index_row = metadata['index_row']
     dec = import_util.safe_column(index_row, 'DECLINATION')
     return dec
-
-def populate_obs_mission_cassini_COVIMS_mission_phase_name_OBS(**kwargs):
-    return helper_cassini_mission_phase_name(**kwargs)
-
-def populate_obs_mission_cassini_COVIMS_sequence_id_OBS(**kwargs):
-    metadata = kwargs['metadata']
-    index_row = metadata['index_row']
-    seqid = index_row['SEQ_ID']
-    return seqid
 
 
 ### OBS_TYPE_IMAGE TABLE ###
@@ -463,6 +394,15 @@ def populate_obs_mission_cassini_COVIMS_spacecraft_clock_count2_OBS(**kwargs):
         sc_cvt = sc1
 
     return sc_cvt
+
+def populate_obs_mission_cassini_COVIMS_mission_phase_name_OBS(**kwargs):
+    return helper_cassini_mission_phase_name(**kwargs)
+
+def populate_obs_mission_cassini_COVIMS_sequence_id_OBS(**kwargs):
+    metadata = kwargs['metadata']
+    index_row = metadata['index_row']
+    seqid = index_row['SEQ_ID']
+    return seqid
 
 
 ################################################################################
